navigation

+++ src/main/java/org/example/CB.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class CB implements PaymentStrategy{
 
-    public String pay(int price) {
+    public String pay(double price) {
         return "Payement effectué en carte bancaire pour un montant de " + price + " €.";
     }
 }
