+++ src/main/java/org/example/PayPal.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class PayPal implements PaymentStrategy{
 
-    public String pay(int price) {
+    public String pay(double price) {
         return "Payement effectué via PayPal pour un montant de " + price + " €.";
     }
 }
