+++ src/main/java/org/example/PaymentStrategy.java
@@ -1,6 +1,6 @@
 package org.example;
 
 public interface PaymentStrategy {
-    String pay(int price);
+    String pay(double price);
 
 }
